\name{exampleADM3}
\alias{exampleADM3}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ An example of using ADM3. }
\description{
Perfroms change point detection on an input file and plots resulting features.
}
\usage{
exampleADM3(outfile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{outfile}{ - location of a report file to be written. }
}
\details{
example.
}
\value{
A report written in file (outfile).
}
\author{ Tomas William Fitzgerald }
\note{ note. }
\examples{
#exampleADM3()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ math }
