\name{plotADM3}
\alias{plotADM3}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ A plotting function. }
\description{
Makes a plot for each detection feature.
}
\usage{
plotADM3(l)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{l}{ - list - output from ADM2. }
}
\details{
Cycles though each detected feature - plotting its surroundings.
}   
\value{
A number of plots.
}
\author{ Tomas William Fitzgerald }
\note{ note. }
\examples{
#plotADM3(l)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ math }
